Benchmarking debugging -Benchmarking now works swiftly -Improved output pipeline -Improved charts aesthetics -Fixed tools' command bugs -Fixed typos

+++ DPAMSA/main.py
@@ -13,7 +13,7 @@ dataset = dataset1
 
 def main():
     config.DEVICE = torch.device(config.DEVICE_NAME)
-    multi_train(dataset, truncate_file=True)
+    multi_train(dataset, truncate_file=False)
 
 
 def output_parameters():
@@ -39,16 +39,27 @@ def multi_train(dataset=dataset, start=0, end=-1, truncate_file=False, model_pat
     output_parameters()
 
     tag = os.path.splitext(dataset.file_name)[0]
-    report_file_name = os.path.join(utils.DPAMSA_REPORTS_PATH, f"{tag}.rpt")
-    csv_file_name = os.path.join(utils.CSV_PATH, "DPAMSA_training", f"{tag}.csv")
+    report_file_name = os.path.join(config.DPAMSA_REPORTS_PATH, f"{tag}.txt")
+    csv_file_name = os.path.join(config.CSV_PATH, "DPAMSA_training", f"{tag}.csv")
 
     if truncate_file:
+        # Se truncate_file è True, tronca i file
         with open(report_file_name, 'w'):
             pass
         with open(csv_file_name, 'w', newline='') as csv_file:
             writer = csv.writer(csv_file)
             writer.writerow(["Dataset Name", "Number of Sequences", "Alignment Length", "SP Score", "Exact Matches",
                              "Column Score"])
+    else:
+        # Se truncate_file è False, crea il file se non esiste e scrivi l'intestazione
+        if not os.path.exists(report_file_name):
+            with open(report_file_name, 'w'):
+                pass
+        if not os.path.exists(csv_file_name):
+            with open(csv_file_name, 'w', newline='') as csv_file:
+                writer = csv.writer(csv_file)
+                writer.writerow(["File Name", "Number of Sequences (QTY)", "Alignment Length (AL)", "Sum of Pairs (SP)",
+                                 "Exact Matches (EM)", "Column Score (CS)"])
 
     if truncate_file:
         with open(report_file_name, 'w'):
@@ -100,14 +111,15 @@ def multi_train(dataset=dataset, start=0, end=-1, truncate_file=False, model_pat
         exact_matches = env.calc_exact_matched()
         column_score = exact_matches / alignment_length
         num_sequences = len(env.aligned)
+        # Crea il report testuale
 
         report = (
-            f"#: {name}\n"
-            f"AL: {alignment_length}\n"
-            f"QTY: {num_sequences}\n"
-            f"SP: {sp_score}\n"
-            f"EM: {exact_matches}\n"
-            f"CS: {column_score}\n"
+            f"File: {name}\n"
+            f"Alignment Length (AL): {alignment_length}\n"
+            f"Number of Sequences (QTY): {num_sequences}\n"
+            f"Sum of Pairs (SP): {sp_score}\n"
+            f"Exact Matches (EM): {exact_matches}\n"
+            f"Column Score (CS): {column_score}:.3f\n"
             f"Alignment:\n{env.get_alignment()}\n\n"
         )
 
@@ -119,8 +131,8 @@ def multi_train(dataset=dataset, start=0, end=-1, truncate_file=False, model_pat
             writer.writerow([name, alignment_length, num_sequences, sp_score, exact_matches, column_score])
 
     print(f"\nOperazione completata con successo.")
-    print(f"Il file di report è stato salvato in: {utils.DPAMSA_REPORTS_PATH}")
-    print(f"Il file CSV è stato salvato in: {utils.CSV_PATH}")
+    print(f"Il file di report è stato salvato in: {config.DPAMSA_REPORTS_PATH}")
+    print(f"Il file CSV è stato salvato in: {config.CSV_PATH}")
 
 
 def train(index):
@@ -171,20 +183,27 @@ def inference(dataset=dataset, start=0, end=-1, model_path='model_3x30', truncat
     output_parameters()
 
     tag = os.path.splitext(dataset.file_name)[0]
-    report_file_name = os.path.join(utils.DPAMSA_REPORTS_PATH, f"{tag}.rpt")
-    csv_file_name = os.path.join(utils.INFERENCE_CSV_PATH, "DPAMSA", f"{tag}.csv")
+    report_file_name = os.path.join(config.DPAMSA_REPORTS_PATH, f"{tag}.txt")
+    csv_file_name = os.path.join(config.DPAMSA_INF_CSV_PATH, f"{tag}_DPAMSA_results.csv")
 
     if truncate_file:
+        # Se truncate_file è True, tronca i file
         with open(report_file_name, 'w'):
             pass
         with open(csv_file_name, 'w', newline='') as csv_file:
             writer = csv.writer(csv_file)
-            writer.writerow(["Dataset Name", "Number of Sequences", "Alignment Length", "SP Score", "Exact Matches",
-                             "Column Score"])
-
-    if truncate_file:
-        with open(report_file_name, 'w'):
-            pass
+            writer.writerow(["File Name", "Number of Sequences (QTY)", "Alignment Length (AL)", "Sum of Pairs (SP)",
+                             "Exact Matches (EM)", "Column Score (CS)"])
+    else:
+        # Se truncate_file è False, crea il file se non esiste e scrivi l'intestazione
+        if not os.path.exists(report_file_name):
+            with open(report_file_name, 'w'):
+                pass
+        if not os.path.exists(csv_file_name):
+            with open(csv_file_name, 'w', newline='') as csv_file:
+                writer = csv.writer(csv_file)
+                writer.writerow(["File Name", "Number of Sequences (QTY)", "Alignment Length (AL)", "Sum of Pairs (SP)",
+                                 "Exact Matches (EM)", "Column Score (CS)"])
 
     datasets_to_process = dataset.datasets[start:end if end != -1 else len(dataset.datasets)]
 
@@ -212,12 +231,12 @@ def inference(dataset=dataset, start=0, end=-1, model_path='model_3x30', truncat
 
         # Crea il report testuale
         report = (
-            f"#: {name}\n"
-            f"AL: {metrics['AL']}\n"
-            f"QTY: {metrics['QTY']}\n"
-            f"SP: {metrics['SP']}\n"
-            f"EM: {metrics['EM']}\n"
-            f"CS: {metrics['CS']}\n"
+            f"File: {name}\n"
+            f"Alignment Length (AL): {metrics['AL']}\n"
+            f"Number of Sequences (QTY): {metrics['QTY']}\n"
+            f"Sum of Pairs (SP): {metrics['SP']}\n"
+            f"Exact Matches (EM): {metrics['EM']}\n"
+            f"Column Score (CS): {metrics['CS']:.3f}\n"
             f"Alignment:\n{env.get_alignment()}\n\n"
         )
 
@@ -231,8 +250,8 @@ def inference(dataset=dataset, start=0, end=-1, model_path='model_3x30', truncat
             writer.writerow([name, metrics['AL'], metrics['QTY'], metrics['SP'], metrics['EM'], metrics['CS']])
 
     print(f"\nOperazione completata con successo.")
-    print(f"Il file di report è stato salvato in: {utils.DPAMSA_REPORTS_PATH}")
-    print(f"Il file CSV è stato salvato in: {utils.CSV_PATH}")
+    print(f"Il file di report è stato salvato in: {config.DPAMSA_REPORTS_PATH}")
+    print(f"Il file CSV è stato salvato in: {config.CSV_PATH}\n\n")
 
 
 if __name__ == "__main__":
